pyproject.toml
[project]
name = "voice-agent-tester"
version = "0.1.0"
description = "A command-line tool to test voice agents."
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "pyppeteer>=1.0.2",
    "pyyaml>=6.0",
    "pydantic>=2.4.2"
]

[project.scripts]
tester = "voice_agent_tester.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
